Why are gradle error logs completely nonsensical

+++ src/main/java/team/lodestar/lodestone/modules/toolkit/sound/SoundPlayer.java
@@ -15,8 +15,8 @@ import java.util.function.*;
  */
 public class SoundPlayer {
 
-    private float minPitch = 1, maxPitch = 1;
-    private float minVolume = 1, maxVolume = 1;
+    private double minPitch = 1, maxPitch = 1;
+    private double minVolume = 1, maxVolume = 1;
     private final SoundEvent soundEvent;
 
     public static SoundPlayer create(DeferredHolder<SoundEvent, SoundEvent> soundEvent) {
@@ -39,21 +39,21 @@ public class SoundPlayer {
         this.soundEvent = soundEvent;
     }
 
-    public SoundPlayer pitch(float pitch) {
+    public SoundPlayer pitch(double pitch) {
         return pitch(pitch, pitch);
     }
 
-    public SoundPlayer pitch(float min, float max) {
+    public SoundPlayer pitch(double min, double max) {
         minPitch = min;
         maxPitch = max;
         return this;
     }
 
-    public SoundPlayer volume(float volume) {
+    public SoundPlayer volume(double volume) {
         return volume(volume, volume);
     }
 
-    public SoundPlayer volume(float min, float max) {
+    public SoundPlayer volume(double min, double max) {
         minVolume = min;
         maxVolume = max;
         return this;
@@ -69,8 +69,8 @@ public class SoundPlayer {
 
     public SoundPlayer play(Level level, Position position, SoundSource source) {
         var random = level.random;
-        float volume = Easing.SINE_IN_OUT.asWeighedRandom(random, minVolume, maxVolume);
-        float pitch = Easing.SINE_IN_OUT.asWeighedRandom(random, minPitch, maxPitch);
+        float volume = (float) Easing.SINE_IN_OUT.asWeighedRandom(random, minVolume, maxVolume);
+        float pitch = (float) Easing.SINE_IN_OUT.asWeighedRandom(random, minPitch, maxPitch);
         level.playSound(null, position.x(), position.y(), position.z(), soundEvent, source, volume, pitch);
         return this;
     }
